requirements.txt
GoogleNews
newspaper3k
openai
streamlit
streamlit_extras